\name{RANKS-package}
\alias{RANKS-package}
\alias{RANKS}
\docType{package}
\title{
\packageTitle{RANKS}
}

\description{
\packageDescription{RANKS}
}

\details{

The DESCRIPTION file:
\packageDESCRIPTION{RANKS}
\packageIndices{RANKS}

}

\author{
\packageAuthor{RANKS}

Maintainer: \packageMaintainer{RANKS}
}
\references{
Re M, Mesiti M, Valentini G: A fast ranking algorithm for predicting gene functions in biomolecular networks.
IEEE ACM Trans Comput Biol Bioinform 2012, 9(6):1812-1818.

Re M, Valentini G: Cancer module genes ranking using kernelized score functions.
BMC Bioinformatics 2012, 13(S14):S3.

Re M, Valentini G: Network-based drug ranking and repositioning with respect to DrugBank therapeutic categories.
IEEE/ACM Trans Comput Biol Bioinform 2013, 10(6):1359-1371. 

G. Valentini, A. Paccanaro, H. Caniza, A. Romero, M. Re: An extensive analysis of disease-gene associations using network integration
and fast kernel-based gene prioritization methods, Artif. Intell. in Med. 61 (2) (2014) 63-78
}

\keyword{ package }
